titanic_ml/requirements.txt
kafka-python
joblib
pandas
scikit-learn
